Split JAX-RS methods from business logic
1. split FHIRResource in half.  Now `FHIRResource` is just the JAXRS
part.  The "doX" methods that serve as a bridge to FHIRPersistence
methods has been moved to FHIRRestHelper.

2. removed http-specific info (uri, http headers, request properties,
and security context) from FHIRPersistenceEvent...the idea is to add
these back in as FHIRRequestContext properties as needed

3. removed PROPNAME_RESOURCE_HELPER from FHIROperationContext...its
already passed directly to the operations as an argument to `doInvoke`

4. removed httpheaders from FHIRNotificationEvent

5. minor whitespace cleanup in DocumentOperation and
HealthcheckOperation (and tests)
