+++ fhir-notification/src/main/java/com/ibm/fhir/notification/FHIRNotificationEvent.java
@@ -6,8 +6,6 @@
 
 package com.ibm.fhir.notification;
 
-import javax.ws.rs.core.HttpHeaders;
-
 import com.ibm.fhir.model.resource.Resource;
 
 public class FHIRNotificationEvent {
@@ -16,7 +14,6 @@ public class FHIRNotificationEvent {
     private String operationType = null;
     private String resourceId = null;
     private Resource resource = null;
-    private HttpHeaders httpHeaders = null;
 
     public FHIRNotificationEvent() {
     }
@@ -61,7 +58,6 @@ public class FHIRNotificationEvent {
                 + ", resourceId=" + getResourceId()
                 + ", location=" + getLocation()
                 + ", lastUpdated=" + getLastUpdated()
-                + ", httpHeaders=" + getHttpHeaders()
                 + "]");
         return sb.toString();
     }
@@ -72,13 +68,5 @@ public class FHIRNotificationEvent {
 
     public void setResource(Resource resource) {
         this.resource = resource;
-    }
-
-    public HttpHeaders getHttpHeaders() {
-        return this.httpHeaders;
-    }
-
-    public void setHttpHeaders(HttpHeaders httpHeaders) {
-        this.httpHeaders = httpHeaders;
     }
 }
