+++ fhir-notification/src/main/java/com/ibm/fhir/notification/FHIRNotificationService.java
@@ -226,7 +226,6 @@ public class FHIRNotificationService implements FHIRPersistenceInterceptor {
             event.setLocation((String) pEvent.getProperty(FHIRPersistenceEvent.PROPNAME_RESOURCE_LOCATION_URI));
             event.setResourceId(resource.getId());
             event.setResource(resource);
-            event.setHttpHeaders(pEvent.getHttpHeaders());
 
             return event;
         } catch (Exception e) {
